+++ src/main/java/kr/codesqaud/cafe/controller/SignInController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import kr.codesqaud.cafe.config.Session;
-import kr.codesqaud.cafe.controller.dto.SignInDTO;
+import kr.codesqaud.cafe.controller.dto.SignInRequest;
 import kr.codesqaud.cafe.controller.dto.user.UserDTO;
 import kr.codesqaud.cafe.service.UserService;
 
@@ -22,19 +22,19 @@ public class SignInController {
 		this.userService = userService;
 	}
 
-	@PostMapping("/user/sign-in-success")
-	public String userSignIn(@ModelAttribute SignInDTO signInDto, HttpSession httpSession) {
-		String id = signInDto.getId();
-		userService.matchPassword(signInDto);
+	@PostMapping("/user/sign-in")
+	public String userSignIn(@ModelAttribute SignInRequest signInRequest, HttpSession httpSession) {
+		String id = signInRequest.getId();
+		userService.matchPassword(signInRequest);
 
 		UserDTO userDto = userService.getUserById(id);
 		Session session = new Session(userDto.getId(), userDto.getNickName());
 
 		httpSession.setAttribute(Session.LOGIN_USER, session);
-		return "redirect:/user/sign-in-success/" + id;
+		return "redirect:/user/sign-in/" + id;
 	}
 
-	@GetMapping("/user/sign-in-success/{Id}")
+	@GetMapping("/user/sign-in/{Id}")
 	public String showSingInSuccessForm(@PathVariable String Id, Model model) {
 		model.addAttribute("user", userService.getUserById(Id));
 		return "user/login_success";
