+++ src/main/resources/templates/fragments/navbar.html
@@ -5,7 +5,7 @@
         <div class="col-md-12">
             <div class="navbar-header">
                 <a href="/" class="navbar-brand">
-                    <img src="/images/KakaoCI.png" />
+                    <img src="/images/KakaoCI.png"/>
                 </a>
             </div>
 
@@ -13,12 +13,14 @@
                 <li class="active"><a href="/user/list">멤버 리스트</a></li>
                 {{#loginUser}}
                 <form method="post" action="/user/sign-out">
-                <li><button class="black-component" type="submit">로그아웃</button></li>
+                    <li>
+                        <button class="black-component" type="submit">로그아웃</button>
+                    </li>
                 </form>
                 <li><a class="black-component" href="/user/profile/{{loginUser.id}}/form" role="button">개인정보수정</a></li>
                 {{/loginUser}}
                 {{^loginUser}}
-                <li><a class="black-component" href="/user/sign-in" role="button">로그인</a></li>
+                <li><a class="black-component" href="/user/sign-in-form" role="button">로그인</a></li>
                 <li><a class="black-component" href="/user/sign-up-form" role="button">회원가입</a></li>
                 {{/loginUser}}
             </ul>
