+++ src/main/resources/templates/user/login.html
@@ -12,10 +12,11 @@
         <header class="post-header">
             <p class="title">로그인</p>
         </header>
-        <form name="question" class="info-container" method="post" action="/user/sign-in-success">
+        <form name="question" class="info-container" method="post" action="/user/sign-in">
             <div class="form-group">
                 <label for="id">아이디</label>
-                <input type="userId" class="form-control" id="id" name="id" placeholder="아이디를 입력하세요" onclick="hideIdErrorMessage(this)">
+                <input type="userId" class="form-control" id="id" name="id" placeholder="아이디를 입력하세요"
+                       onclick="hideIdErrorMessage(this)">
             </div>
             {{#id-error}}
             <div>
@@ -24,7 +25,8 @@
             {{/id-error}}
             <div class="form-group">
                 <label for="password">비밀번호</label>
-                <input type="password" class="form-control" id="password" name="password" placeholder="비밀번호를 입력하세요" onclick="hidePasswordErrorMessage(this)">
+                <input type="password" class="form-control" id="password" name="password" placeholder="비밀번호를 입력하세요"
+                       onclick="hidePasswordErrorMessage(this)">
             </div>
             {{#password-error}}
             <div>
